projects: Merge both ad738x projects
There are two projects for ad738x:

ad738x-st: Using the new project format and supporting stm32 only
ad738x_fmcz: Using the old project format and supporting xilinx.

Merge the two projects into a single one.

Some fixes for the xilinx project were added during the move
like added pwm and clkgen init parameters.
