--- projects/ad738x-st/Makefile
@@ -1,8 +1,0 @@
-# Select the example you want to enable by chossing y for enabling and n for disabling
-IIO_EXAMPLE = y
-
-include ../../tools/scripts/generic_variables.mk
-
-include src.mk
-
-include ../../tools/scripts/generic.mk
