--- projects/ad738x-st/builds.json
@@ -1,7 +1,0 @@
-{
-  "stm32": {
-    "iio_example": {
-      "flags" : "IIO_EXAMPLE=y HARDWARE=sdp-ck1z.ioc"
-    }
-  }
-}
