+++ projects/ad738x_fmcz/Makefile
@@ -1,3 +1,7 @@
+# Select the example you want to enable by chossing y for enabling and n for disabling
+BASIC_EXAMPLE = n
+IIO_EXAMPLE = y
+
 include ../../tools/scripts/generic_variables.mk
 
 include src.mk
