+++ projects/ad738x_fmcz/src/common/common_data.h
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   iio_example.c
- *   @brief  IIO example header for eval-ad738x project
+ *   @file   common_data.h
+ *   @brief  Defines common data to be used by eval-ad738x examples.
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
@@ -37,64 +37,22 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
+#ifndef __COMMON_DATA_H__
+#define __COMMON_DATA_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include <stdio.h>
-
-#include "iio_example.h"
-#include "common_data.h"
-#include "iio.h"
+#include "platform_includes.h"
 #include "ad738x.h"
-#include "iio_ad738x.h"
-#include "no_os_util.h"
-#include "no_os_gpio.h"
-#include "no_os_print_log.h"
-#include "iio_app.h"
 
 /******************************************************************************/
-/************************ Functions Declarations ******************************/
+/********************** Macros and Constants Definitions **********************/
+/******************************************************************************/
+#define ADC_DDR_BASEADDR	in_buff
+extern uint8_t in_buff[];
+extern struct no_os_uart_init_param ad738x_uart_ip;
+extern struct ad738x_init_param ad738x_init_param;
+extern struct no_os_uart_desc *uart_desc;
 
-/**
- * @brief IIO example main execution.
- *
- * @return ret - Result of the example execution. If working correctly, will
- *               execute continuously the while(1) loop and will not return.
- */
-int iio_example_main()
-{
-	struct ad738x_iio_dev *dev;
-	struct iio_app_init_param app_init_param = {0};
-	struct iio_app_desc *app;
-	struct iio_app_device iio_devices[] = {
-		{ .name = "ad738x" },
-	};
-	int ret;
-
-	ret = ad738x_iio_init(&dev, &ad738x_ip);
-	if (ret)
-		return ret;
-
-	iio_devices[0].dev = dev;
-	iio_devices[0].dev_descriptor = dev->iio_dev;
-	app_init_param.devices = iio_devices;
-	app_init_param.nb_devices = NO_OS_ARRAY_SIZE(iio_devices);
-	app_init_param.uart_init_params = ad738x_uart_ip;
-
-	ret = iio_app_init(&app, app_init_param);
-	if (ret) {
-		ad738x_iio_remove(dev);
-		return ret;
-	}
-
-	ret = iio_app_run(app);
-	if (ret)
-		pr_info("Error: iio_app_run: %d\n", ret);
-
-	iio_app_remove(app);
-
-	ad738x_iio_remove(dev);
-
-	return ret;
-}
+#endif /* __COMMON_DATA_H__ */
