+++ projects/ad738x_fmcz/src/platform/platform_includes.h
@@ -1,9 +1,10 @@
 /***************************************************************************//**
- *   @file   app_config.h
- *   @brief  Config file of AD738x-FMCZ.
- *   @author Antoniu Miclaus (antoniu.miclaus@analog.com)
+ *   @file   platform_includes.h
+ *   @brief  Includes for used platforms used by eval-ad738x project.
+ *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
- * Copyright 2020(c) Analog Devices, Inc.
+ * Copyright 2024(c) Analog Devices, Inc.
+ * Copyright 2024(c) BayLibre, SAS.
  *
  * All rights reserved.
  *
@@ -36,13 +37,20 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef APP_CONFIG_H_
-#define APP_CONFIG_H_
+#ifndef __PLATFORM_INCLUDES_H__
+#define __PLATFORM_INCLUDES_H__
 
 /******************************************************************************/
-/********************** Macros and Constants Definitions **********************/
+/***************************** Include Files **********************************/
 /******************************************************************************/
-
-#define SPI_ENGINE_OFFLOAD_EXAMPLE
-
+#ifdef XILINX_PLATFORM
+#include "xilinx/parameters.h"
+#elif defined STM32_PLATFORM
+#include "stm32/parameters.h"
 #endif
+
+#ifdef IIO_SUPPORT
+#include "iio_app.h"
+#endif
+
+#endif /* __PLATFORM_INCLUDES_H__ */
