+++ projects/ad738x_fmcz/src/platform/stm32/platform_src.mk
