+++ projects/ad738x_fmcz/src/platform/xilinx/parameters.c
@@ -1,10 +1,10 @@
 /***************************************************************************//**
- *   @file   common_data.c
- *   @brief  Defines common data to be used by eval-ad738x examples.
+ *   @file   parameters.c
+ *   @brief  Definition of xilinx platform data used by ad463x_fmcz project.
+ *   @author Antoniu Miclaus (antoniu.miclaus@analog.com)
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
- * Copyright 2024(c) BayLibre, SAS.
  *
  * All rights reserved.
  *
@@ -41,31 +41,17 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "common_data.h"
-#include "main.h"
-/******************************************************************************/
-/********************** Macros and Constants Definitions **********************/
-/******************************************************************************/
-struct no_os_uart_init_param ad738x_uart_ip = {
-	.device_id = UART_DEVICE_ID,
-	.irq_id = UART_IRQ_ID,
-	.asynchronous_rx = true,
-	.baud_rate = UART_BAUDRATE,
-	.size = NO_OS_UART_CS_8,
-	.parity = NO_OS_UART_PAR_NO,
-	.stop = NO_OS_UART_STOP_1_BIT,
-	.extra = UART_EXTRA,
-	.platform_ops = UART_OPS,
+#include "parameters.h"
+
+struct xil_uart_init_param uart_extra_ip = {
+#ifdef XPAR_XUARTLITE_NUM_INSTANCES
+	.type = UART_PL,
+#else
+	.type = UART_PS,
+	.irq_id = UART_IRQ_ID
+#endif
 };
 
-struct no_os_spi_init_param ad738x_spi_ip = {
-	.device_id = SPI_DEVICE_ID,
-	.max_speed_hz = SPI_BAUDRATE,
-	.bit_order = NO_OS_SPI_BIT_ORDER_MSB_FIRST,
-	.mode = NO_OS_SPI_MODE_2,
-	.platform_ops = SPI_OPS,
-	.chip_select = SPI_CS,
-	.extra = SPI_EXTRA,
-};
 
-struct ad738x_init_param ad738x_ip;
+
+
