+++ projects/ad738x_fmcz/src/platform/xilinx/platform_src.mk
@@ -1,0 +1,19 @@
+
+SRCS +=	$(PLATFORM_DRIVERS)/xilinx_axi_io.c \
+	$(PLATFORM_DRIVERS)/xilinx_spi.c \
+	$(PLATFORM_DRIVERS)/xilinx_delay.c \
+	$(PLATFORM_DRIVERS)/xilinx_uart.c \
+	$(PLATFORM_DRIVERS)/xilinx_irq.c \
+	$(PLATFORM_DRIVERS)/xilinx_gpio.c
+
+SRCS +=	$(DRIVERS)/axi_core/axi_dmac/axi_dmac.c \
+	$(DRIVERS)/axi_core/spi_engine/spi_engine.c \
+	$(DRIVERS)/axi_core/axi_pwmgen/axi_pwm.c \
+	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.c
+
+INCS += $(DRIVERS)/axi_core/axi_dmac/axi_dmac.h \
+	$(DRIVERS)/axi_core/spi_engine/spi_engine.h \
+	$(DRIVERS)/axi_core/spi_engine/spi_engine_private.h \
+	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.h \
+	$(DRIVERS)/axi_core/axi_pwmgen/axi_pwm_extra.h \
+	$(PLATFORM_DRIVERS)/xilinx_gpio.h
